+++ pyiron_workflow/snippets/singleton.py
